feat: unify RoundTrip implementation in auth providers
